Able to register and login

+++ lib/cadeau_compas/wishlist/queries/queries.sql
@@ -29,21 +29,32 @@ FROM
   wishlists W
   LEFT JOIN wishlist_products WP ON WP.wishlist_id = W.id
   LEFT JOIN products P ON P.id = WP.product_id
+WHERE
+  W.user_id = :user_id :: text :: uuid
 ORDER BY
   W.inserted_at DESC,
   WP.inserted_at DESC;
 
 -- name: insert_to_wishlist
-WITH inserted_row AS (
+WITH authorized_wishlist AS (
+  SELECT
+    w.id
+  FROM
+    wishlists AS w
+  WHERE
+    w.id = :wishlist_id :: text :: uuid
+    AND w.user_id = :user_id :: text :: uuid
+),
+inserted_row AS (
   INSERT INTO
     wishlist_products (wishlist_id, product_id, inserted_at, updated_at)
-  VALUES
-    (
-      :wishlist_id :: text :: uuid,
-      :product_id :: text :: uuid,
-      now(),
-      now()
-    ) RETURNING product_id
+  SELECT
+    AW.id,
+    :product_id :: text :: uuid,
+    now(),
+    now()
+  FROM
+    authorized_wishlist AW RETURNING product_id
 )
 SELECT
   P.id :: text,
@@ -58,11 +69,20 @@ FROM
 DELETE FROM
   wishlists
 WHERE
-  id = :wishlist_id :: text :: uuid;
+  id = :wishlist_id :: text :: uuid
+  AND user_id = :user_id :: text :: uuid;
 
 -- name: delete_product_from_list
 DELETE FROM
   wishlist_products
 WHERE
-  wishlist_id = :wishlist_id :: text :: uuid
-  AND product_id = :product_id :: text :: uuid;
+  product_id = :product_id :: text :: uuid
+  AND wishlist_id IN (
+    SELECT
+      w.id
+    FROM
+      wishlists AS w
+    WHERE
+      w.id = :wishlist_id :: text :: uuid
+      AND w.user_id = :user_id :: text :: uuid
+  );
